P1C3.1 transtipage

+++ P1/P1C3.1/Program.cs
@@ -2,7 +2,7 @@
 int denominateur = 4;
 
 // TODO : modifier l'instruction ci-dessous pour que la réponse prenne une valeur décimale
-int reponse = numerateur / denominateur;
+decimal reponse = numerateur / (decimal)denominateur;
 
 // Afficher le résultat 
 Console.WriteLine(numerateur + " / " + denominateur + " = " + reponse);
